+++ src/main/java/org/folio/cataloging/business/cataloguing/authority/AuthorityNameTitleHeadingTag.java
@@ -25,7 +25,6 @@ public class AuthorityNameTitleHeadingTag extends AuthorityHeadingTag {
 	/**
 	 * Class constructor
 	 *
-	 * @param d
 	 * @since 1.0
 	 */
 	public AuthorityNameTitleHeadingTag() {
@@ -35,7 +34,7 @@ public class AuthorityNameTitleHeadingTag extends AuthorityHeadingTag {
 	/* (non-Javadoc)
 	 * @see TagInterface#getCategory()
 	 */
-	public short getCategory() {
+	public int getCategory() {
 		return 11;
 	}
