+++ src/main/java/org/folio/cataloging/business/cataloguing/bibliographic/BibliographicTagImpl.java
@@ -86,7 +86,7 @@ public class BibliographicTagImpl extends TagImpl {
 	 * 
 	 * @see TagImpl#getValidEditableSubfields(short)
 	 */
-	public Set getValidEditableSubfields(short category) {
+	public Set getValidEditableSubfields(int category) {
 		Set set = new TreeSet(new SubfieldCodeComparator());
 		switch (category) {
 		case 2:
@@ -112,7 +112,7 @@ public class BibliographicTagImpl extends TagImpl {
 
 	@Override
 	public Correlation getCorrelation(String tagNumber, char indicator1,
-			char indicator2, short category) throws DataAccessException {
+			char indicator2, int category) throws DataAccessException {
 		return daoCorrelation.getBibliographicCorrelation(tagNumber,
 				indicator1, indicator2, category);
 	}
